Extract `filter_records` to filter.py
